requirements.txt
azure.functions
paramiko
pandas
pytz
openpyxl
msal
requests